pyproject.toml
[tool.poetry]
name = "dw"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
requests = "^2.32.5"
python-dotenv = "^1.1.1"
psycopg2 = "^2.9.10"
django = "^5.2.5"
redis = "^6.4.0"
djangorestframework-simplejwt = "^5.5.1"
pillow = "^11.3.0"
django-cors-headers = "^4.7.0"
docker = "^7.1.0"
drf-yasg = "^1.21.10"
serializers = "^0.2.4"
djangorestframework = "^3.16.1"
django-filter = "^25.1"
pytest = "^8.4.1"
pytest-django = "^4.11.1"
psycopg2-binary = "^2.9.10"
pytest-cov = "^6.2.1"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.3.0"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.17.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
